MySQL Provider: ForeignKeyBuilder
This commit adds a ForeignKeyBuilder for MySQL, implementing the IKeyDeclBuilder<T> interaface. The builder is
relatively straightforward, creating a single SQL snippet that is valid for a MySQL provider. It enforces a 64-character
limit on the length of a key name. The builder allows 'SET DEFAULT' for on-delete and on-update actions, even though
these are likely to be rejected by the MySQL engine for several database implementations.