MySQL Provider: ConstraintBuilder
This commit adds a ConstraintBuilder for MySQL, implementing the IConstraintBuilder<T> interface. I don't love the
design here, but it is greatly complicated by the fact that maximum length constraints are realized in the Field's type
rather than as a CHECK. To deal with this, the builder produces an abstract type that is effectively a union: it is
either a maximum length constraint or a regular constraint, and it will ultimately be the TableBuilder's responsibility
to arbitrate between themmm. The builder enforces a 64-character limit on the length of a constraint name, even though
Kvasir currently has no mechanism for supplying them.