MySQL Provider: FieldBuilder
This commit adds a FieldBuilder for MySQL, implementing the IFieldBuilder<T> interface. The builder is relatively
straightforward, though it does produce an intermediate (rather than a SQLSnippet directly) in order to arbitrate
between TEXT and VARCHAR. The intermediate accepts a maximum length, at which point the type symbol converts from the
former to the latter; when built, the intermediate produces the SQLSnippet. The builder enforces a 64-character limit
on the length of a field name.