MySQL Provider: TableBuilder
This commit adds a TableBuilder for MySQL, implementing the ITableBuilder<...> interface. The builder is relatively
straightforward, tracking the various pieces and then composing them into a single CREATE TABLE IF EXISTS statement. The
main piece of logic is in applying maximum-length constraints to the appropriate Fields to convert TEXT to VARCHAR(N).
The builder enforces a 64-character limit on the length of a table name.