Insert Pre-Defined Rows on Table Creation
This commit updates the logic for the CreateTable function of the Transactor to insert all rows for all Pre-Defined
Entities at the same time. To do this, the pre-defined instances are now stored as a list of objects on the
PrincipalTableDef; for regular Entities, this list is empty. All pre-defined instances are inserted as a single
transaction after the transaction to create the tables is commited; this way, any database providers that don't support
concurrent transactions or mixing of DDL and DML are satisfied.