When Deleting a Row of an Associative Relation during an Entity Update, Pass Only Primary Key Fields
This commit fixes a bug whereby we were passing too many Fields to the ICommand.Delete function when handling deleted
rows of an associative Relation while updating an Entity. Previously, in such a circumstance, we passed the entire row
to the Delete function; for Maps and OrderedLists, this included more than just the key/index, which is enough to
uniquely identify the row to be deleted. This commit fixes that so the minimal amount of data is forwarded. In addition,
this commit update the documentation of the ICommand.Delete function to make it clear exactly what is expected to be
passed.