Support System.DateOnly as a CLR Type
This commit adds support for CLR properties backed by System.DateOnly. This is a counterpart to System.DateOnly and is
handled basically the same way: defaults are strings parsed into the CLR type and comparison constraints are allowed. I
have renamed the DateTimeFieldDescriptor to CalendarFieldDescriptor to reflect the fact that it no longer only
represents System.DateTime.

The MySQL provider code has been updated as well, and it didn't take much. We use the DATE keyword for both the column
type and the literal identifier.