Bug Fix: MySQL DateTime Literal Keyword is TIMESTAMP not DATETIME
This commit fixes a bug in the rendering of DateTime literals for the MySQL provider. Previously, the keyword identifier
used was DATETIME, but this was wrong according to the official documentation. Instead, it should be TIMESTAMP. Note
that DateOnly literals (using the DATE keyword) are unaffected. (The column type identifier for DateTime fields is also
unaffected.)