Make VARCHAR(255) the Data Type for Text Fields in MySQL
This commit changes, yet again, the behavior for Text-type Fields in the MySQL provider. Previously, they defaulted to
being represented with LONGTEXT and would become VARCHAR(N) when a [Check.LengthIsAtMost] constraint was applied. This
was problematic, though, because most of the time people want VARCHAR(N), and requiring users to annotate everything is
a pain. After considering a bunch of alternatives (including a [Blob] annotation and a custom C# type), I decided that
65,535 characters should be just plenty.

So, now Text-type Fields default to being represented with VARCHAR(255). The presence of a [Check.LengthIsAtMost]
constraint can reduce or increase the 255-character limit up to 65,535. As a consequence, I have also fixed a bug that
previously allowed overlong length specifications for VARCHARs.