Localization Data Structure
This commit adds the Localization data structure, which is a map-like implementation of the IRelation interface designed
to handle loalized data (e.g. text-per-language, date-per-calendar, measurement-per-system). Each Localization is
initialized with a key that identifies the semantic being localized; it then has a series of locale-to-value mappings
exposed through the standard Dictionary API. The connection type exposed to Kvasir is a three-element tuple.