Simplify Relation Interfaces
This commit removes a number of interfaces explicitly listed for the four Relation containers, since those interfaces
were already present through others. For example, IEnumerable<T> is implied by IList<T>, and therefore does not need to
be listed in the file directly. There is no impact to correctness, only to brevity.