Translation of Localizations
This commit introduces Localizations, a new first-class property type that Kvasir can handle. A Localization
is any type derived from the Localization<K, L, V> generic, which in turn implements the brand new
ILocalization interface. Localizations are designed to support different values under different circumstances
for a single semantic, such as different languages' translations of text or different systems' units of
measurement.

The "key" of a Localization shows up in the Principal or Relation Table that "owns" the Localization, while
the locale-value pairs are aggregated into a Localization Table. All instances of a Localization Type share
a single Localization Table, so the keys must be unique-per-type, while the locales are necessarily unique
per-key. Derived Localization types cannot have their own properties included in the data model.

Annotations may be applied to a Localization's key in the Principal or Relation Table, but those constraints
do not propagate to the Localization Table. Data Converters are not supported, however. Localizations may
appear nested in Relations, but not in other Localizations; Relations cannot be nested in Localizations.

At this time, only the Schema-layer translation of Localizations is supported. Extraction and Reconstitution
will be implemented in future commits.