Localization Data Structure
This commit adds the Localization data structure to the new Localization namespace. This data structure is essentially a
map, correlating locale to values. Each Localization is identified by a localization key (which defines the semantic
being localized) and then exposes its mapping in two ways: operator[] and as an IReadOnlyDictionary through a dedicated
property. Internally, the mapping is stored as a RelationMap, so it automatically performs change tracking.