Translation of Localizations
This commit implements the entirety of the Translation layer for Localizations. Localizations translate much like a
regular Entity, though the fields are implicit: the Key field is based on a bona fide property, while the Locale and Key
fields are based on the underlying Relation. A Localization reifies as its Key when on an Entity or a Relation. Most
annotations are allowed, but data conversions are not permitted. Localizations can be empty, and therefore they do not
result in Foreign Keys.

To support Localizations, a number of changes have been made:

   * A new this[] API has been added to the Translator, since the return defs are subtley different
   * Localizations are identified as their own category
   * The LocalizationKeyFieldGroup represents a Localization on a foreign table
   * The LocalizationTracker is used to contextualize a Localization, whose translation is deferred
   * Contexts track their reference explicitly, and they can be 'cleared' to avoid inducing cycles
   * The permissibility of various kinds of fields is now controlled by a traits enum